av1: Introduce alm_step for having smaller step-size
This commit introduces --alm_step, so we can experiments
with 0.25 changes like 1.25, 1.50, 1.75. we are having as
additional parameter for two reasons
1. Precision and flexibiilty
2. Adaption, as the current CLI args do not support input
of double/float variables, we did with this method. So now
we can have single scirpt with 1..10 values to have
iteration of values.
By default it is 0, either you can pass 0 or do not include
in the call.

Example: --alm_k 1 --alm_step=2, gives us 1.50.

Also cleans up 02a7363e0 with comment and styling issues.